+++ components_graphic.go
@@ -1,0 +1,191 @@
+package katsu2d
+
+import (
+	"image"
+	"image/color"
+
+	ebimath "github.com/edwinsyarief/ebi-math"
+	"github.com/hajimehoshi/ebiten/v2/text/v2"
+	"golang.org/x/text/language"
+)
+
+type SpriteComponent struct {
+	TextureID int
+	SrcX      float32
+	SrcY      float32
+	SrcW      float32
+	SrcH      float32
+	DstW      float32
+	DstH      float32
+	Color     color.RGBA
+	Opacity   float32
+}
+
+func NewSpriteComponent(textureID, width, height int) *SpriteComponent {
+	return &SpriteComponent{
+		TextureID: textureID,
+		DstW:      float32(width),
+		DstH:      float32(height),
+		Color:     color.RGBA{255, 255, 255, 255},
+		Opacity:   1.0,
+	}
+}
+
+func (self *SpriteComponent) GetSourceRect(textureWidth, textureHeight float32) (x, y, w, h float32) {
+	x, y = self.SrcX, self.SrcY
+	w, h = self.SrcW, self.SrcH
+	if w == 0 || h == 0 {
+		w, h = textureWidth, textureHeight
+	}
+	return
+}
+
+func (self *SpriteComponent) GetDestSize(sourceWidth, sourceHeight float32) (w, h float32) {
+	w, h = self.DstW, self.DstH
+	if w == 0 {
+		w = sourceWidth
+	}
+	if h == 0 {
+		h = sourceHeight
+	}
+	return
+}
+
+type AnimMode int
+
+const (
+	AnimLoop AnimMode = iota
+	AnimOnce
+	AnimBoomerang
+)
+
+type AnimationComponent struct {
+	Frames    []image.Rectangle
+	Speed     float64
+	Elapsed   float64
+	Current   int
+	Mode      AnimMode
+	Direction bool
+	Active    bool
+}
+
+type TextAlignment int
+
+const (
+	TextAlignmentBottomRight TextAlignment = iota
+	TextAlignmentMiddleRight
+	TextAlignmentTopRight
+	TextAlignmentBottomCenter
+	TextAlignmentMiddleCenter
+	TextAlignmentTopCenter
+	TextAlignmentBottomLeft
+	TextAlignmentMiddleLeft
+	TextAlignmentTopLeft
+)
+
+var alignmentOffsets = map[TextAlignment]func(w, h float64) (float64, float64){
+	TextAlignmentTopRight:     func(w, h float64) (float64, float64) { return 0, 0 },
+	TextAlignmentMiddleRight:  func(w, h float64) (float64, float64) { return 0, -h / 2 },
+	TextAlignmentBottomRight:  func(w, h float64) (float64, float64) { return 0, -h },
+	TextAlignmentTopCenter:    func(w, h float64) (float64, float64) { return -w / 2, 0 },
+	TextAlignmentMiddleCenter: func(w, h float64) (float64, float64) { return -w / 2, -h / 2 },
+	TextAlignmentBottomCenter: func(w, h float64) (float64, float64) { return -w / 2, -h },
+	TextAlignmentTopLeft:      func(w, h float64) (float64, float64) { return 0, 0 },
+	TextAlignmentMiddleLeft:   func(w, h float64) (float64, float64) { return 0, -h / 2 },
+	TextAlignmentBottomLeft:   func(w, h float64) (float64, float64) { return 0, -h },
+}
+
+type TextComponent struct {
+	Alignment         TextAlignment
+	Caption           string
+	Size, lineSpacing float64
+	Color             color.RGBA
+	fontFace          *text.GoTextFace
+	cachedWidth       float64
+	cachedHeight      float64
+	cachedText        string
+}
+
+func NewTextComponent(source *text.GoTextFaceSource, caption string, size float64, col color.RGBA) *TextComponent {
+	fontFace := &text.GoTextFace{
+		Source:    source,
+		Direction: text.DirectionLeftToRight,
+		Size:      size,
+		Language:  language.English,
+	}
+	result := &TextComponent{
+		Caption:  caption,
+		Size:     size,
+		Color:    col,
+		fontFace: fontFace,
+	}
+	result.updateCache()
+	return result
+}
+
+func (self *TextComponent) updateCache() {
+	if self.cachedText != self.Caption {
+		self.cachedWidth, self.cachedHeight = text.Measure(self.Caption, self.fontFace, self.lineSpacing)
+		self.cachedText = self.Caption
+	}
+}
+
+func (self *TextComponent) LineSpacing() float64 {
+	return self.lineSpacing
+}
+
+func (self *TextComponent) SetLineSpacing(spacing float64) *TextComponent {
+	self.lineSpacing = spacing
+	self.updateCache()
+	return self
+}
+
+func (self *TextComponent) SetAlignment(alignment TextAlignment) *TextComponent {
+	self.Alignment = alignment
+	return self
+}
+
+func (self *TextComponent) GetOffset() (float64, float64) {
+	if offsetFunc, ok := alignmentOffsets[self.Alignment]; ok {
+		offsetX, offsetY := offsetFunc(self.cachedWidth, self.cachedHeight)
+		return offsetX, offsetY
+	}
+	return 0, 0
+}
+
+func (self *TextComponent) SetText(text string) *TextComponent {
+	if self.Caption != text {
+		self.Caption = text
+		self.updateCache()
+	}
+	return self
+}
+
+func (self *TextComponent) SetFontFace(fontFace *text.GoTextFace) *TextComponent {
+	self.fontFace = fontFace
+	self.updateCache()
+	return self
+}
+
+func (self *TextComponent) FontFace() *text.GoTextFace {
+	return self.fontFace
+}
+
+func (self *TextComponent) SetSize(size float64) *TextComponent {
+	self.fontFace.Size = size
+	self.updateCache()
+	return self
+}
+
+func (self *TextComponent) SetColor(color color.RGBA) *TextComponent {
+	self.Color = color
+	return self
+}
+
+func (self *TextComponent) SetOpacity(opacity float64) *TextComponent {
+	val := ebimath.Min(ebimath.Max(opacity, 0.0), 1.0)
+	col := self.Color
+	col.A = uint8(255 * val)
+	self.SetColor(col)
+	return self
+}
